gettext macos updates

+++ gettext/0.20.1/macos/install.sh
@@ -9,6 +9,10 @@ if [ ! -d "$HOME/programs" ]; then
 	mkdir "$HOME/programs"
 fi
 
+if [ ! -d "$HOME/logs" ]; then
+	mkdir "$HOME/logs"
+fi
+
 if [ ! -d "$HOME/sources/$FOLDER_NAME" ]; then
 	mkdir "$HOME/sources/$FOLDER_NAME"
 fi
@@ -17,24 +21,45 @@ if [ ! -d "$HOME/programs/$FOLDER_NAME" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME"
 fi
 
+if [ ! -d "$HOME/logs/$FOLDER_NAME" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME"
+fi
+
+if [ ! -d "$HOME/logs/$FOLDER_NAME/$VERSION" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME/$VERSION"
+fi
+
 if [ ! -d "$HOME/programs/$FOLDER_NAME/$VERSION" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME/$VERSION"
 
 	cd $HOME/sources/$FOLDER_NAME
 
+	printf "${bold}${yellow}Installing $FOLDER_NAME${clear}\n"
+
 	export CC=gcc
-	wget "https://ftp.gnu.org/gnu/gettext/gettext-$VERSION.tar.gz"
-	tar -xvf "gettext-$VERSION.tar.gz"
+	printf "\t${bold}${green}Downloading source code${clear}\n"
+	wget -q --show-progress "https://ftp.gnu.org/gnu/gettext/gettext-$VERSION.tar.gz"
+	printf "\t${bold}${green}Extracting source code${clear}\n"
+	tar -xf "gettext-$VERSION.tar.gz"
 	mv "gettext-$VERSION" $VERSION
 	cd $VERSION
-	./configure --prefix=$HOME/programs/$FOLDER_NAME/$VERSION
-	make
-	sudo make install
+	printf "\t${bold}${green}Configuring${clear}\n"
+	./configure --help > $HOME/logs/$FOLDER_NAME/$VERSION/configureHelp.txt 2>&1
+	./configure --prefix=$HOME/programs/$FOLDER_NAME/$VERSION > $HOME/logs/$FOLDER_NAME/$VERSION/configureOutput.txt 2>&1
+	printf "\t${bold}${green}Making${clear}\n"
+	make > $HOME/logs/$FOLDER_NAME/$VERSION/makeOutput.txt 2>&1
+	printf "\t${bold}${green}Installing${clear}\n"
+	echo $USER_PASSWORD | sudo -S -p '' make install > $HOME/logs/$FOLDER_NAME/$VERSION/installOutput.txt 2>&1
 
-	cd $HOME/programs/$FOLDER_NAME/$VERSION
-	sudo chown -R $(whoami) .
+	if [ -e "$HOME/programs/$FOLDER_NAME/$VERSION/bin/gettext" ]; then
+		cd $HOME/programs/$FOLDER_NAME/$VERSION
+		echo $USER_PASSWORD | sudo -S -p '' chown -R $(whoami) .
 
-	cd $HOME/sources/$FOLDER_NAME
-	rm -rf $VERSION
-	rm "gettext-$VERSION.tar.gz"
+		printf "\t${bold}${green}Clearing${clear}\n"
+		cd $HOME/sources/$FOLDER_NAME
+		rm -rf $VERSION
+		rm "gettext-$VERSION.tar.gz"
+	fi
 fi
+
+cd $HOME/install-files
